student-face-sorter-mvp/requirements.txt
insightface==0.7.3
onnxruntime==1.22.1
opencv-python==4.10.0.84
numpy==2.1.1
scikit-learn==1.5.2
python-dotenv==1.0.1
pandas==2.2.3
